requirements.txt
Flask
torch
open_clip_torch
Pillow
pandas